requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
idna==3.3
joblib==1.1.0
nltk==3.7
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.15
requests==2.27.1
six==1.16.0
tornado==6.1
tqdm==4.63.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.9